Fix double key press bug when cycling through the associated connections
Remove incorrect +1 offsets in connection cycling logic that were causing
user to need to press 'n' twice to move to the next connection. The
offsets were remnants from when connections were stashed back into the
connections vector differently. (now just it pushes to the end of the vector,
used to insert it back at the index it was taken out)